+++ include/overworld/Engine/Common/Urdf/VisualActor.h
@@ -16,8 +16,8 @@ namespace owds {
     explicit VisualActor(const std::vector<owds::Shape>& visual_shapes);
 
     ~VisualActor() noexcept override {};
-    void setup(const std::array<float, 3>& position,
-               const std::array<float, 4>& orientation) override;
+    void setup(const std::array<double, 3>& position,
+               const std::array<double, 4>& orientation) override;
 
     void remove() override {}
 
@@ -32,18 +32,18 @@ namespace owds {
      * @param position Absolute position.
      * @param orientation Absolute rotation, in radians.
      */
-    void setPositionAndOrientation(const std::array<float, 3>& position, const std::array<float, 4>& orientation) override;
-    void setVelocity(const std::array<float, 3>& linear_velocity, const std::array<float, 3>& angular_velocity) override;
+    void setPositionAndOrientation(const std::array<double, 3>& position, const std::array<double, 4>& orientation) override;
+    void setVelocity(const std::array<double, 3>& linear_velocity, const std::array<double, 3>& angular_velocity) override;
 
     AABB_t getAABB() override;
     AABB_t getLocalAABB() override;
 
     std::array<float, 16> getModelMatrix() const override;
-    std::pair<std::array<float, 3>, std::array<float, 4>> getPositionAndOrientation() const override;
+    std::pair<std::array<double, 3>, std::array<double, 4>> getPositionAndOrientation() const override;
 
   private:
-    std::array<float, 3> position_;
-    std::array<float, 4> orientation_;
+    std::array<double, 3> position_;
+    std::array<double, 4> orientation_;
   };
 } // namespace owds
